+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 ### Changed
 * moved datasets provider to its own directory and leverage the registration function
+* Moved provider options validation to ProviderRegistration
 
 ### Fixed
 * path construction when no host or id parameter